Bug 1718271 - wasm: Make rtt depth optional and use in validation. r=lth
Rtts are allowed to specify an optional depth based on which type
code they are encoded with. We currently always require the depth,
but don't validate it or use it anywhere. This commit allows omitting
the rtt depth and actually validating it correctly when it is
specified.

An AbstractRttCode is added and used in ValType as the canonical
type code for rtts, as they can now be decoded from either
TypeCode::Rtt or TypeCode::RttWithDepth.

A tricky bug was uncovered through testing this where BlockType
and ResultType would silently truncate ValType (which is uintptr_t
sized) to a uint32_t for their immediate forms.

Differential Revision: https://phabricator.services.mozilla.com/D118844

+++ js/src/wasm/WasmTypes.h
@@ -267,7 +267,7 @@ class ResultType {
     InvalidKind = Tagged::PointerKind2,
   };
 
-  ResultType(Kind kind, uint32_t imm) : tagged_(Tagged::Kind(kind), imm) {}
+  ResultType(Kind kind, uintptr_t imm) : tagged_(Tagged::Kind(kind), imm) {}
   explicit ResultType(const ValTypeVector* ptr)
       : tagged_(Tagged::Kind(VectorKind), ptr) {}
 
@@ -286,7 +286,7 @@ class ResultType {
  public:
   ResultType() : tagged_(Tagged::Kind(InvalidKind), nullptr) {}
 
-  static ResultType Empty() { return ResultType(EmptyKind, uint32_t(0)); }
+  static ResultType Empty() { return ResultType(EmptyKind, uintptr_t(0)); }
   static ResultType Single(ValType vt) {
     return ResultType(SingleKind, vt.bitsUnsafe());
   }
@@ -380,7 +380,7 @@ class BlockType {
     FuncResultsKind = Tagged::PointerKind2
   };
 
-  BlockType(Kind kind, uint32_t imm) : tagged_(Tagged::Kind(kind), imm) {}
+  BlockType(Kind kind, uintptr_t imm) : tagged_(Tagged::Kind(kind), imm) {}
   BlockType(Kind kind, const FuncType& type)
       : tagged_(Tagged::Kind(kind), &type) {}
 
@@ -398,7 +398,7 @@ class BlockType {
                 PackedTypeCode::invalid().bits()) {}
 
   static BlockType VoidToVoid() {
-    return BlockType(VoidToVoidKind, uint32_t(0));
+    return BlockType(VoidToVoidKind, uintptr_t(0));
   }
   static BlockType VoidToSingle(ValType vt) {
     return BlockType(VoidToSingleKind, vt.bitsUnsafe());
